app/requirements.txt
Flask
tensorflow-cpu
numpy
sklearn